Improvements to building "my" AESHASH.
AES wants to set up a key schedule at server startup based on
the original 128 bit key. This means our seed buffer is larger
than needs to be initialized, or shown in this case. So we add
various build primitives to make it work.

"my" AESHASH is loosely based on the one from Go. I would love
to have it be *directly* based on the one from Go, but my C/ASM
chops aren't quite there yet. :-(